Build fixes for Haiku.
git-svn-id: svn://localhost/APlayer/trunk@2 2acc5306-6531-47f8-909e-9acfd2f9568f

+++ APlayer/Players/Fred/Fred.cpp
@@ -745,7 +745,7 @@ ap_result Fred::LoadFinal(PFile *file, PString &errorStr)
 		}
 
 		// Allocate space to hold the inst/note commands
-		instNoteSize  = abs(instOffset - instNoteOffset);
+		instNoteSize  = abs((int)(instOffset - instNoteOffset));
 		instNoteTable = new uint8[instNoteSize];
 		if (instNoteTable == NULL)
 		{
